moved bim360/docs endpoint to new docm.py file
